refactor(install): split fetch pipeline
